apps/compliance-checker/pyproject.toml
[project]
name = "compliance-checker"
version = "0.0.1"
description = "Multi-framework building compliance checker MCP server (BBL + WKB)"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.27.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0",
    "aioboto3>=13.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.4",
    "mypy>=1.13.0",
    "types-PyYAML>=6.0",
    "types-beautifulsoup4>=4.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/compliance_checker"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "SIM", "TCH", "RUF"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["compliance_checker"]
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["aioboto3.*", "botocore.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
